# doc/swagger/AoL-Teacher.yaml
...
openapi: 3.0.0
info:
  title: AoL Teacher Interface
  description: |-
    AoL Teacher API specification.

  contact:
    email: sudhagar@abovecloud9.ai
  version: 0.0.9
...
servers:
  - url: https://dev.api.ac9ai.com/v1   # development
  - url: https://api.ac9ai.com/v1       # production
...
paths:
...
  /media/metadata:
    get:
      tags:
        - media
      summary: Get media metadata
      description: Retrieves metadata for both quotes & images
      responses:
        '200':
          description: Metadata successfully retrieved
          content:
            application/json:
              schema:
                type: object
                properties:
                  quote:
                    $ref: '#/components/schemas/MediaMetadata'
                  image:
                    $ref: '#/components/schemas/MediaMetadata'
        '400':
          description: Bad request. 
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
        - bearer: []
  /media/metadata/{type}:
    get:
      tags:
        - media
      summary: Get metadata
      description: Retrieve metadata for quotes or images
      parameters:
        - name: type
          in: path
          required: true
          description: The type of media to fetch metadata for (e.g., "quote" or "image")
          schema:
            type: string
            enum:
              - quote
              - image
      responses:
        '200':
          description: Metadata successfully retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MediaMetadata'
        '400':
          description: Bad request. 
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
        - bearer: []
...
  /push-notify/register:
    post:
      tags:
        - push-notify
      summary: Register push notification information to GLAD
      description: ''
      operationId: storePushNotifyInfo
      requestBody:
        description: Push notification information
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PushNotifyInfo'
        required: true
      responses:
        '200':
          description: successful operation
        '500':
          description: Server error occurred
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
        - bearer: []
components:
  schemas:
...
    PushNotifyInfo:
      type: object
      properties:
        pushToken:
          type: string
        # Note: Identifier that Apple/Google sends when token is invalidated
        revokeID:
          type: string
        appVersion:
          type: string
          # Note: Format to be finalized
          example: 2024.12.1
        deviceInfo:     # For analytics
          type: object
        platformInfo:   # For analytics
          type: object

    # HTTP status code indicates whether the request is a success or failure
    # In case of failre, error object will contain the details. In case of success
    # schema defined in the spec above will be returned.
    ErrorResponse:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: integer
              format: int32
            message:
              type: string
            details:
              type: string
            traceID:
              type: string
...
    MediaMetadata:
      type: object
      properties:
        version:
          type: integer
          description: Version of the metadata
          format: int64
          example: 1
        lastUpdated:
          type: string
          format: date-time
          description: Last updated timestamp
        total:
          type: integer
          description: Total number of items (quotes or images) in the system
        url:
          type: string
          description: URL to fetch actual content or detailed metadata
  securitySchemes:
    bearer:
      type: apiKey
      name: Bearer
      in: header